Revert "Make picoruby-mbedtls cipher module portable across platforms."
This reverts commit 1ed733d234d8a469fdbd5944dd33c0faf90526b0.
