+++ mrbgems/picoruby-mbedtls/src/mruby/cipher.c
@@ -4,7 +4,6 @@
 #include "mruby/array.h"
 #include "mruby/data.h"
 #include "mruby/class.h"
-#include "cipher.h"
 
 static void
 mrb_cipher_free(mrb_state *mrb, void *ptr)
@@ -24,19 +23,32 @@ mrb_mbedtls_cipher_initialize(mrb_state *mrb, mrb_value self)
 
   mbedtls_cipher_type_t cipher_type;
   uint8_t key_len, iv_len;
-  MbedTLS_cipher_type_key_iv_len(cipher_name, &cipher_type, &key_len, &iv_len);
+  mbedtls_cipher_type_key_iv_len(cipher_name, &cipher_type, &key_len, &iv_len);
   if (cipher_type == MBEDTLS_CIPHER_NONE) {
     mrb_raise(mrb, E_ARGUMENT_ERROR, "unsupported cipher suite");
   }
 
-  unsigned char * instance_data = mrb_malloc(mrb, MbedTLS_cipher_instance_size());
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_malloc(mrb, sizeof(cipher_instance_t));
   DATA_PTR(self) = instance_data;
   DATA_TYPE(self) = &mrb_cipher_type;
-  int ret = MbedTLS_cipher_new(instance_data, cipher_type, MBEDTLS_OPERATION_NONE, key_len, iv_len);
-  if (ret == CIPHER_NEW_BAD_INPUT_DATA) {
-    mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_setup failed (bad input data)");
-  } else if (ret == CIPHER_NEW_FAILED_TO_SETUP) {
-    mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_setup failed");
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+  mbedtls_cipher_init(ctx);
+  instance_data->cipher_type = cipher_type;
+  instance_data->operation = MBEDTLS_OPERATION_NONE;
+  instance_data->key_len = key_len;
+  instance_data->iv_len = iv_len;
+  instance_data->key_set = false;
+  instance_data->iv_set = false;
+
+  const mbedtls_cipher_info_t *cipher_info = mbedtls_cipher_info_from_type(cipher_type);
+  int ret;
+  ret = mbedtls_cipher_setup(ctx, cipher_info);
+  if (ret != 0) {
+    if (ret == MBEDTLS_ERR_CIPHER_BAD_INPUT_DATA) {
+      mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_setup failed (bad input data)");
+    } else {
+      mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_setup failed");
+    }
   }
   return self;
 }
@@ -46,8 +58,7 @@ mrb_mbedtls_cipher_ciphers(mrb_state *mrb, mrb_value self)
 {
   mrb_value ret = mrb_ary_new_capa(mrb, CIPHER_SUITES_COUNT);
   for (int i = 0; i < CIPHER_SUITES_COUNT; i++) {
-    const char *cipher_name = MbedTLS_cipher_cipher_name(i);
-    mrb_value str = mrb_str_new_cstr(mrb, cipher_name);
+    mrb_value str = mrb_str_new_cstr(mrb, cipher_suites[i].name);
     mrb_ary_set(mrb, ret, i, str);
   }
   return ret;
@@ -56,22 +67,24 @@ mrb_mbedtls_cipher_ciphers(mrb_state *mrb, mrb_value self)
 static mrb_value
 mrb_mbedtls_cipher_encrypt(mrb_state *mrb, mrb_value self)
 {
-  MbedTLS_cipher_encrypt(mrb_data_get_ptr(mrb, self, &mrb_cipher_type));
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  instance_data->operation = MBEDTLS_ENCRYPT;
   return self;
 }
 
 static mrb_value
 mrb_mbedtls_cipher_decrypt(mrb_state *mrb, mrb_value self)
 {
-  MbedTLS_cipher_decrypt(mrb_data_get_ptr(mrb, self, &mrb_cipher_type));
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  instance_data->operation = MBEDTLS_DECRYPT;
   return self;
 }
 
 static mrb_value
 mrb_mbedtls_cipher_key_len(mrb_state *mrb, mrb_value self)
 {
-  uint8_t key_len = MbedTLS_cipher_key_len(mrb_data_get_ptr(mrb, self, &mrb_cipher_type));
-  return mrb_fixnum_value(key_len);
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  return mrb_fixnum_value(instance_data->key_len);
 }
 
 static mrb_value
@@ -80,20 +93,33 @@ mrb_mbedtls_cipher_key_eq(mrb_state *mrb, mrb_value self)
   mrb_value key;
   mrb_get_args(mrb, "S", &key);
 
-  int ret = MbedTLS_cipher_key_eq(mrb_data_get_ptr(mrb, self, &mrb_cipher_type), RSTRING_PTR(key), RSTRING_LEN(key));
-  if (ret == CIPHER_KEY_EQ_DOUBLE) {
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  if (instance_data->key_set) {
     mrb_warn(mrb, "key should be set once per instance, ignoring\n");
     return key;
-  } else if (ret == CIPHER_KEY_EQ_OPERATION_NOT_SET) {
+  }
+  if (instance_data->operation == MBEDTLS_OPERATION_NONE) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "operation is not set");
-  } else if (ret == CIPHER_KEY_EQ_INVALID_KEY_LENGTH) {
+  }
+  if (RSTRING_LEN(key) != instance_data->key_len) {
     mrb_raise(mrb, E_ARGUMENT_ERROR, "key length is invalid");
-  } else if (ret == CIPHER_KEY_EQ_FAILED_TO_SET_KEY) {
-    mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_setkey failed");
-  } else if (ret == CIPHER_KEY_EQ_FAILED_TO_SET_PADDING_MODE) {
-    mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_set_padding_mode failed");
   }
 
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  int ret;
+  ret = mbedtls_cipher_setkey(ctx, (const unsigned char *)RSTRING_PTR(key), RSTRING_LEN(key) * 8, instance_data->operation); /* last arg is keybits */
+  if (ret != 0) {
+    mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_setkey failed");
+  }
+  if (mbedtls_cipher_is_cbc(instance_data->cipher_type)) {
+    ret = mbedtls_cipher_set_padding_mode(ctx, MBEDTLS_PADDING_PKCS7);
+    if (ret != 0) {
+      mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_set_padding_mode failed");
+    }
+  }
+
+  instance_data->key_set = true;
   //mrb_incref(&v[0]);
   //mrb_incref(&key);
   return key;
@@ -102,8 +128,8 @@ mrb_mbedtls_cipher_key_eq(mrb_state *mrb, mrb_value self)
 static mrb_value
 mrb_mbedtls_cipher_iv_len(mrb_state *mrb, mrb_value self)
 {
-  uint8_t iv_len = MbedTLS_cipher_iv_len(mrb_data_get_ptr(mrb, self, &mrb_cipher_type));
-  return mrb_fixnum_value(iv_len);
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  return mrb_fixnum_value(instance_data->iv_len);
 }
 
 static mrb_value
@@ -112,20 +138,31 @@ mrb_mbedtls_cipher_iv_eq(mrb_state *mrb, mrb_value self)
   mrb_value iv;
   mrb_get_args(mrb, "S", &iv);
 
-  int ret = MbedTLS_cipher_iv_eq(mrb_data_get_ptr(mrb, self, &mrb_cipher_type), RSTRING_PTR(iv), RSTRING_LEN(iv));
-  if (ret == CIPHER_IV_EQ_DOUBLE) {
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  if (instance_data->iv_set) {
     mrb_warn(mrb, "iv should be set once per instance, ignoring\n");
     return iv;
-  } else if (ret == CIPHER_IV_EQ_OPERATION_NOT_SET) {
+  }
+  if (instance_data->operation == MBEDTLS_OPERATION_NONE) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "operation is not set");
-  } else if (ret == CIPHER_IV_EQ_INVALID_IV_LENGTH) {
+  }
+  if (RSTRING_LEN(iv) != instance_data->iv_len) {
     mrb_raise(mrb, E_ARGUMENT_ERROR, "iv length is invalid");
-  } else if (ret == CIPHER_IV_EQ_FAILED_TO_SET_IV) {
+  }
+
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  int ret;
+  ret = mbedtls_cipher_set_iv(ctx, (const unsigned char *)RSTRING_PTR(iv), RSTRING_LEN(iv));
+  if (ret != 0) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_set_iv failed");
-  } else if (ret == CIPHER_IV_EQ_FAILED_TO_RESET) {
+  }
+  ret = mbedtls_cipher_reset(ctx);
+  if (ret != 0) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_reset failed");
   }
 
+  instance_data->iv_set = true;
   //mrb_incref(&v[0]);
   //mrb_incref(&iv);
   return iv;
@@ -137,8 +174,12 @@ mrb_mbedtls_cipher_update_ad(mrb_state *mrb, mrb_value self)
   mrb_value input;
   mrb_get_args(mrb, "S", &input);
 
-  int ret = MbedTLS_cipher_update_ad(mrb_data_get_ptr(mrb, self, &mrb_cipher_type), (const char *)RSTRING_PTR(input), RSTRING_LEN(input));
-  if (ret == CIPHER_UPDATE_AD_FAILED) {
+  int ret;
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_update_ad(ctx, (const unsigned char *)RSTRING_PTR(input), RSTRING_LEN(input));
+  if (ret != 0) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_update_ad failed");
   }
 
@@ -154,9 +195,13 @@ mrb_mbedtls_cipher_update(mrb_state *mrb, mrb_value self)
 
   size_t out_len = RSTRING_LEN(input) + 16;
   unsigned char* output = (unsigned char *)mrb_malloc(mrb, out_len);
+  int ret;
 
-  int ret = MbedTLS_cipher_update(mrb_data_get_ptr(mrb, self, &mrb_cipher_type), (const char *)RSTRING_PTR(input), RSTRING_LEN(input), output, &out_len);
-  if (ret == CIPHER_UPDATE_FAILED) {
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_update(ctx, (const unsigned char *)RSTRING_PTR(input), RSTRING_LEN(input), output, &out_len);
+  if (ret != 0) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_update failed");
   }
   mrb_value ret_value = mrb_str_new(mrb, (const char *)output, (mrb_int)out_len);
@@ -170,8 +215,13 @@ mrb_mbedtls_cipher_finish(mrb_state *mrb, mrb_value self)
 {
   size_t out_len = 16;
   unsigned char output[out_len];
-  int ret = MbedTLS_cipher_finish(mrb_data_get_ptr(mrb, self, &mrb_cipher_type), output, &out_len);
-  if (ret == CIPHER_FINISH_FAILED) {
+  int ret;
+
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_finish(ctx, output, &out_len);
+  if (ret != 0) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_finish failed");
   }
   mrb_value ret_value = mrb_str_new(mrb, (const char *)output, (mrb_int)out_len);
@@ -182,11 +232,15 @@ mrb_mbedtls_cipher_finish(mrb_state *mrb, mrb_value self)
 static mrb_value
 mrb_mbedtls_cipher_write_tag(mrb_state *mrb, mrb_value self)
 {
+  int ret;
   size_t tag_len = 16;
   unsigned char tag[tag_len];
 
-  int ret = MbedTLS_cipher_write_tag(mrb_data_get_ptr(mrb, self, &mrb_cipher_type), (const char *)tag, tag_len);
-  if (ret == CIPHER_WRITE_TAG_FAILED) {
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_write_tag(ctx, tag, tag_len);
+  if (ret != 0) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_write_tag failed");
   }
 
@@ -201,11 +255,15 @@ mrb_mbedtls_cipher_check_tag(mrb_state *mrb, mrb_value self)
   mrb_value input;
   mrb_get_args(mrb, "S", &input);
 
-  int ret = MbedTLS_cipher_check_tag(mrb_data_get_ptr(mrb, self, &mrb_cipher_type), (const char *)RSTRING_PTR(input), RSTRING_LEN(input));
-  if (ret == CIPHER_CHECK_TAG_AUTH_FAILED) {
-    // mrb_incref(&self);
+  int ret;
+  cipher_instance_t *instance_data = (cipher_instance_t *)mrb_data_get_ptr(mrb, self, &mrb_cipher_type);
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_check_tag(ctx, (const unsigned char *)RSTRING_PTR(input), RSTRING_LEN(input));
+  if (ret == MBEDTLS_ERR_CIPHER_AUTH_FAILED) {
+    //mrb_incref(&v[0]);
     return mrb_false_value();
-  } else if (ret == CIPHER_CHECK_TAG_FAILED) {
+  } else if (ret != 0) {
     mrb_raise(mrb, E_RUNTIME_ERROR, "mbedtls_cipher_check_tag failed");
   }
   //mrb_incref(&v[0]);
