+++ mrbgems/picoruby-mbedtls/src/mrubyc/cipher.c
@@ -1,5 +1,4 @@
 #include <mrubyc.h>
-#include "cipher.h"
 
 static void
 c_mbedtls_cipher_new(mrbc_vm *vm, mrbc_value *v, int argc)
@@ -20,19 +19,32 @@ c_mbedtls_cipher_new(mrbc_vm *vm, mrbc_value *v, int argc)
   }
   mbedtls_cipher_type_t cipher_type;
   uint8_t key_len, iv_len;
-  MbedTLS_cipher_type_key_iv_len(cipher_name, &cipher_type, &key_len, &iv_len);
+  mbedtls_cipher_type_key_iv_len(cipher_name, &cipher_type, &key_len, &iv_len);
   if (cipher_type == MBEDTLS_CIPHER_NONE) {
     mrbc_raise(vm, MRBC_CLASS(ArgumentError), "unsupported cipher suite");
     return;
   }
 
-  mrbc_value self = mrbc_instance_new(vm, v->cls, MbedTLS_cipher_instance_size());
-  int ret = MbedTLS_cipher_new(self.instance->data, cipher_type, MBEDTLS_OPERATION_NONE, key_len, iv_len);
-  if (ret == CIPHER_NEW_BAD_INPUT_DATA) {
-    mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_setup failed (bad input data)");
-    return;
-  } else if (ret == CIPHER_NEW_FAILED_TO_SETUP) {
-    mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_setup failed");
+  mrbc_value self = mrbc_instance_new(vm, v->cls, sizeof(cipher_instance_t));
+  cipher_instance_t *instance_data = (cipher_instance_t *)self.instance->data;
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+  mbedtls_cipher_init(ctx);
+  instance_data->cipher_type = cipher_type;
+  instance_data->operation = MBEDTLS_OPERATION_NONE;
+  instance_data->key_len = key_len;
+  instance_data->iv_len = iv_len;
+  instance_data->key_set = false;
+  instance_data->iv_set = false;
+
+  const mbedtls_cipher_info_t *cipher_info = mbedtls_cipher_info_from_type(cipher_type);
+  int ret;
+  ret = mbedtls_cipher_setup(ctx, cipher_info);
+  if (ret != 0) {
+    if (ret == MBEDTLS_ERR_CIPHER_BAD_INPUT_DATA) {
+      mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_setup failed (bad input data)");
+    } else {
+      mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_setup failed");
+    }
     return;
   }
   SET_RETURN(self);
@@ -43,8 +55,7 @@ c_mbedtls_cipher_ciphers(mrbc_vm *vm, mrbc_value *v, int argc)
 {
   mrbc_value ret = mrbc_array_new(vm, CIPHER_SUITES_COUNT);
   for (int i = 0; i < CIPHER_SUITES_COUNT; i++) {
-    const char *cipher_name = MbedTLS_cipher_cipher_name(i);
-    mrbc_value str = mrbc_string_new_cstr(vm, cipher_name);
+    mrbc_value str = mrbc_string_new_cstr(vm, cipher_suites[i].name);
     mrbc_array_set(&ret, i, &str);
   }
   SET_RETURN(ret);
@@ -53,20 +64,22 @@ c_mbedtls_cipher_ciphers(mrbc_vm *vm, mrbc_value *v, int argc)
 static void
 c_mbedtls_cipher_encrypt(mrbc_vm *vm, mrbc_value *v, int argc)
 {
-  MbedTLS_cipher_encrypt(v->instance->data);
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  instance_data->operation = MBEDTLS_ENCRYPT;
 }
 
 static void
 c_mbedtls_cipher_decrypt(mrbc_vm *vm, mrbc_value *v, int argc)
 {
-  MbedTLS_cipher_decrypt(v->instance->data);
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  instance_data->operation = MBEDTLS_DECRYPT;
 }
 
 static void
 c_mbedtls_cipher_key_len(mrbc_vm *vm, mrbc_value *v, int argc)
 {
-  uint8_t key_len = MbedTLS_cipher_key_len(v->instance->data);
-  SET_INT_RETURN(key_len);
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  SET_INT_RETURN(instance_data->key_len);
 }
 
 static void
@@ -82,25 +95,37 @@ c_mbedtls_cipher_key_eq(mrbc_vm *vm, mrbc_value *v, int argc)
     return;
   }
 
-  int ret = MbedTLS_cipher_key_eq(v->instance->data, (const char *)key.string->data, key.string->size);
-  if (ret == CIPHER_KEY_EQ_DOUBLE) {
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  if (instance_data->key_set) {
     console_printf("[WARN] key should be set once per instance, ignoring\n");
-    SET_RETURN(key);
     return;
-  } else if (ret == CIPHER_KEY_EQ_OPERATION_NOT_SET) {
+  }
+  if (instance_data->operation == MBEDTLS_OPERATION_NONE) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "operation is not set");
     return;
-  } else if (ret == CIPHER_KEY_EQ_INVALID_KEY_LENGTH) {
+  }
+  if (key.string->size != instance_data->key_len) {
     mrbc_raise(vm, MRBC_CLASS(ArgumentError), "key length is invalid");
-    return;
-  } else if (ret == CIPHER_KEY_EQ_FAILED_TO_SET_KEY) {
-    mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_setkey failed");
-    return;
-  } else if (ret == CIPHER_KEY_EQ_FAILED_TO_SET_PADDING_MODE) {
-    mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_set_padding_mode failed");
     return;
   }
 
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  int ret;
+  ret = mbedtls_cipher_setkey(ctx, key.string->data, key.string->size * 8, instance_data->operation); /* last arg is keybits */
+  if (ret != 0) {
+    mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_setkey failed");
+    return;
+  }
+  if (mbedtls_cipher_is_cbc(instance_data->cipher_type)) {
+    ret = mbedtls_cipher_set_padding_mode(ctx, MBEDTLS_PADDING_PKCS7);
+    if (ret != 0) {
+      mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_set_padding_mode failed");
+      return;
+    }
+  }
+
+  instance_data->key_set = true;
   mrbc_incref(&v[0]);
   mrbc_incref(&key);
   SET_RETURN(key);
@@ -109,8 +134,8 @@ c_mbedtls_cipher_key_eq(mrbc_vm *vm, mrbc_value *v, int argc)
 static void
 c_mbedtls_cipher_iv_len(mrbc_vm *vm, mrbc_value *v, int argc)
 {
-  uint8_t iv_len = MbedTLS_cipher_iv_len(v->instance->data);
-  SET_INT_RETURN(iv_len);
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  SET_INT_RETURN(instance_data->iv_len);
 }
 
 static void
@@ -126,25 +151,35 @@ c_mbedtls_cipher_iv_eq(mrbc_vm *vm, mrbc_value *v, int argc)
     return;
   }
 
-  int ret = MbedTLS_cipher_iv_eq(v->instance->data, (const char *)iv.string->data, iv.string->size);
-  if (ret == CIPHER_IV_EQ_DOUBLE) {
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  if (instance_data->iv_set) {
     console_printf("[WARN] iv should be set once per instance, ignoring\n");
-    SET_RETURN(iv);
     return;
-  } else if (ret == CIPHER_IV_EQ_OPERATION_NOT_SET) {
+  }
+  if (instance_data->operation == MBEDTLS_OPERATION_NONE) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "operation is not set");
     return;
-  } else if (ret == CIPHER_IV_EQ_INVALID_IV_LENGTH) {
+  }
+  if (iv.string->size != instance_data->iv_len) {
     mrbc_raise(vm, MRBC_CLASS(ArgumentError), "iv length is invalid");
     return;
-  } else if (ret == CIPHER_IV_EQ_FAILED_TO_SET_IV) {
+  }
+
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  int ret;
+  ret = mbedtls_cipher_set_iv(ctx, iv.string->data, iv.string->size);
+  if (ret != 0) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_set_iv failed");
     return;
-  } else if (ret == CIPHER_IV_EQ_FAILED_TO_RESET) {
+  }
+  ret = mbedtls_cipher_reset(ctx);
+  if (ret != 0) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_reset failed");
     return;
   }
 
+  instance_data->iv_set = true;
   mrbc_incref(&v[0]);
   mrbc_incref(&iv);
   SET_RETURN(iv);
@@ -163,8 +198,12 @@ c_mbedtls_cipher_update_ad(mrbc_vm *vm, mrbc_value *v, int argc)
     return;
   }
 
-  int ret = MbedTLS_cipher_update_ad(v->instance->data, (char *)input.string->data, input.string->size);
-  if (ret == CIPHER_UPDATE_AD_FAILED) {
+  int ret;
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_update_ad(ctx, input.string->data, input.string->size);
+  if (ret != 0) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_update_ad failed");
     return;
   }
@@ -188,9 +227,13 @@ c_mbedtls_cipher_update(mrbc_vm *vm, mrbc_value *v, int argc)
 
   unsigned char* output = mrbc_alloc(vm, input.string->size + 16); // need at least input length + block size
   size_t out_len = input.string->size + 16;
+  int ret;
 
-  int ret = MbedTLS_cipher_update(v->instance->data, (char *)input.string->data, input.string->size, output, &out_len);
-  if (ret == CIPHER_UPDATE_FAILED) {
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_update(ctx, input.string->data, input.string->size, output, &out_len);
+  if (ret != 0) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_update failed");
     return;
   }
@@ -205,8 +248,12 @@ c_mbedtls_cipher_finish(mrbc_vm *vm, mrbc_value *v, int argc)
 {
   unsigned char* output = mrbc_alloc(vm, 16); // need at least block size
   size_t out_len = 16;
+  int ret;
 
-  int ret = MbedTLS_cipher_finish(v->instance->data, output, &out_len);
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_finish(ctx, output, &out_len);
   if (ret != 0) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_finish failed");
     return;
@@ -220,11 +267,15 @@ c_mbedtls_cipher_finish(mrbc_vm *vm, mrbc_value *v, int argc)
 static void
 c_mbedtls_cipher_write_tag(mrbc_vm *vm, mrbc_value *v, int argc)
 {
+  int ret;
   size_t tag_len = 16;
   unsigned char* tag = mrbc_alloc(vm, tag_len); // tag size is 16 for AES-*-GCM and ChaCha20-Poly1305
+                                                //
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
 
-  int ret = MbedTLS_cipher_write_tag(v->instance->data, (const char *)tag, tag_len);
-  if (ret == CIPHER_WRITE_TAG_FAILED) {
+  ret = mbedtls_cipher_write_tag(ctx, tag, tag_len);
+  if (ret != 0) {
     mrbc_raise(vm, MRBC_CLASS(RuntimeError), "mbedtls_cipher_write_tag failed");
     return;
   }
@@ -248,8 +299,12 @@ c_mbedtls_cipher_check_tag(mrbc_vm *vm, mrbc_value *v, int argc)
     return;
   }
 
-  int ret = MbedTLS_cipher_check_tag(v->instance->data, (const char *)input.string->data, input.string->size);
-  if (ret == CIPHER_CHECK_TAG_FAILED) {
+  int ret;
+  cipher_instance_t *instance_data = (cipher_instance_t *)v->instance->data;
+  mbedtls_cipher_context_t *ctx = &instance_data->ctx;
+
+  ret = mbedtls_cipher_check_tag(ctx, input.string->data, input.string->size);
+  if (ret == MBEDTLS_ERR_CIPHER_AUTH_FAILED) {
     mrbc_incref(&v[0]);
     SET_BOOL_RETURN(MRBC_TT_FALSE);
   } else if (ret != 0) {
